refactor: extract and move FocusUnderline and NewFocusUnderline
