+++ src/components/OakLink/OakLink.tsx
@@ -12,11 +12,12 @@ import { OmitKeepDiscriminated } from "../../utils/generics";
 import { box } from "../SharedComponents/Box";
 import { HTMLAnchorProps } from "../SharedComponents/Button/common";
 
-import FocusUnderline, { focusUnderlineStyles } from "./FocusUnderline";
+import FocusUnderline, {
+  focusUnderlineStyles,
+} from "@/components/SharedComponents/FocusUnderline";
 import NewFocusUnderline, {
   newFocusUnderlineStyles,
-} from "./NewFocusUnderline";
-
+} from "@/components/SharedComponents/NewFocusUnderline";
 import flex from "@/styles/utils/flex";
 import Svg from "@/components/SharedComponents/Svg";
 import { FlexProps } from "@/components/SharedComponents/Flex";
