+++ src/components/SharedComponents/Checkbox/Checkbox.tsx
@@ -9,7 +9,7 @@ import spacing, { SpacingProps } from "@/styles/utils/spacing";
 import getColorByLocation from "@/styles/themeHelpers/getColorByLocation";
 import getColorByName from "@/styles/themeHelpers/getColorByName";
 import getFontFamily from "@/styles/themeHelpers/getFontFamily";
-import FocusUnderline from "@/components/OakLink/FocusUnderline";
+import FocusUnderline from "@/components/SharedComponents/FocusUnderline";
 import FieldError from "@/components/SharedComponents/FieldError";
 import { ZIndex } from "@/styles/utils/zIndex";
 import Span from "@/components/SharedComponents/Typography/Span";
