+++ src/components/SharedComponents/FocusUnderline/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./FocusUnderline";
+export { focusUnderlineStyles } from "./FocusUnderline";
