+++ src/components/SharedComponents/NewFocusUnderline/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./NewFocusUnderline";
+export { newFocusUnderlineStyles } from "./NewFocusUnderline";
